Move GetShippedSettingsDirectory to helper
